feat(hippy-react-web): color transform
